docs: Updated documentation template
