+++ stories/Icon/Icon.story.mdx
@@ -1,0 +1,79 @@
+import { Meta, Story } from '@storybook/addon-docs';
+
+import Header from '../shared/Header';
+import { ExampleIcons, ExampleIncorrectIcons } from './Icon';
+import Catalogue from './Catalogue';
+
+<Meta 
+  title="Foundation/Icons" 
+  parameters={{
+    previewTabs: { 
+      'storybook/docs/panel': null, 
+      canvas: {
+        hidden: true,
+      },
+    },
+  }}
+/>
+
+<Header title="Icons">
+    Quantum use Material icons, beautifully crafted symbols for common actions
+    and items. Download on desktop to use them in your digital products for
+    Android, iOS, and web.
+</Header>
+
+## Icons
+
+It's structured with these ones:
+
+- [Usage](#usage)
+- [Catalog](#catalog)
+
+## Importing Icon
+
+```js
+import { Icon } from '@catho/quantum';
+```
+
+### <a id="usage"></a>Usage
+
+You need to import the [typography](?path=/docs/foundation-typography--page) to include icon fonts.
+
+<ExampleIcons />
+
+```js
+<Icon name="info" size="small" />,
+<Icon name="info" />,
+<Icon name="info" size="large" />,
+<Icon name="clear" />
+<Icon name="search" />
+<Icon name="error" />
+<Icon name="visibility_off" />
+<Icon name="visibility" />
+<Icon name="check_box" />
+<Icon name="emoji_people" skin={colors.primary['900']} />
+<Icon name="emoji_people" skin={colors.secondary['900']} />
+<Icon name="emoji_people" skin={colors.primary['700']} />
+<Icon name="emoji_people" skin={colors.warning['900']} />
+<Icon name="emoji_people" skin={colors.error['300']} />
+<Icon name="emoji_people" skin={colors.error['300']} size="large" />
+```
+
+## Unavailable Icons
+
+If the string passed in the `Icon` is unavailable in the style guide
+catalog, it is going to display only the string wrapped in a `span`
+element and no icon will render.
+
+<ExampleIncorrectIcons />
+
+```js
+<Icon name="bike" />
+<Icon name="emoji_smile" skin={colors.primary['900']} />
+```
+
+### <a id="catalog"></a>Catalog
+
+## Available Icons
+
+<Catalogue/>
